# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Navalhex API
  description: Documentação dos contratos da API REST do Navalhex (SaaS Multi-tenant para Barbearias).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento Local
paths:
...
  /api/tenants/{slug}/treatments:
    post:
      tags:
        - Tratamentos & Serviços
      summary: Cadastrar novo serviço na barbearia
      description: Cadastra um serviço/tratamento. Apenas o dono autenticado da barbearia tem permissão.
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: Slug da barbearia
          schema:
            type: string
            example: "navalhex-prime"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTreatmentRequest'
      responses:
        '201':
          description: Serviço cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
              example:
                status: 201
                message: "Serviço criado com sucesso"
                data: null
                errors: null
        '400':
          description: Barbearia não encontrada, campos inválidos ou usuário não é o dono
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: Não autorizado (token ausente ou inválido)

    get:
      tags:
        - Tratamentos & Serviços
      summary: Listar todos os serviços da barbearia (Público)
      description: Retorna a lista de tratamentos e serviços ativos da barbearia a partir do slug.
      parameters:
        - name: slug
          in: path
          required: true
          description: Slug da barbearia
          schema:
            type: string
            example: "navalhex-prime"
      responses:
        '200':
          description: Lista de serviços retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListTreatments'
              example:
                status: 200
                message: "Serviço encontrado com sucesso"
                data:
                  - id: "b2d3e4f5-6a7b-8c9d-0e1f-2a3b4c5d6e7f"
                    title: "Corte Masculino Degradê"
                    description: "Corte com tesoura e máquina."
                    price: 45.00
                    durationMinutes: 35
                errors: null
        '400':
          description: Barbearia não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Envie o token JWT no formato `Bearer <token>`
  schemas:
...
    ApiResponseVoid:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: "Operação realizada com sucesso"
        data:
          type: object
          nullable: true
          default: null
        errors:
          type: object
          nullable: true
          default: null
...
    ApiResponseError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Erro de validação ou regra de negócio"
        data:
          type: object
          nullable: true
          default: null
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            email: "O email deve ser válido"
    CreateTreatmentRequest:
      type: object
      required:
        - title
        - price
        - durationMinutes
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 30
          example: "Corte Masculino Degradê"
        description:
          type: string
          minLength: 10
          maxLength: 255
          example: "Corte moderno com tesoura e máquina."
        price:
          type: number
          format: double
          example: 45.00
        durationMinutes:
          type: integer
          example: 35
    ResponseTreatmentData:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b2d3e4f5-6a7b-8c9d-0e1f-2a3b4c5d6e7f"
        title:
          type: string
          example: "Corte Masculino Degradê"
        description:
          type: string
          example: "Corte moderno com tesoura e máquina."
        price:
          type: number
          format: double
          example: 45.00
        durationMinutes:
          type: integer
          example: 35
    ApiResponseListTreatments:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "Serviço encontrado com sucesso"
        data:
          type: array
          items:
            $ref: '#/components/schemas/ResponseTreatmentData'
        errors:
          type: object
          nullable: true
          default: null